move ChangingProgressbar into its own file, remove COMPONENT_* variables from being dynamically fetched from package.json
